📝 : clarify backlog paths in journey table
